refactor to be more c l e a n
