+++ src/components/Schedule/Constants.js
@@ -1,0 +1,10 @@
+export const HOUR_HEIGHT = 150
+export const EVENT_GAP = 5
+export const EVENT_WIDTH = 400
+
+export const EVENT_TYPES = {
+  "main": ["MAIN", "#FFFFFF"],
+  "minievents": ["ACTIVITY", "#FFE27A"],
+  "workshop": ["WORKSHOP", "#31E0E0"],
+  "notices": ["NOTICE", "#9EFF7C"],
+}
