+++ src/components/Schedule/index.js
@@ -1,0 +1,2 @@
+import Schedule from './Schedule'
+export default Schedule
