+++ src/containers/Schedule.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react'
-import Schedule from '../components/Schedule'
+import Schedule from '../components/Schedule/'
 import { H1 } from '../components/Typography';
 import { db } from '../utility/firebase'
 import { DB_COLLECTION, DB_HACKATHON, DAYOF_COLLECTION } from '../utility/Constants'
