+++ src/pages/Charcuterie.js
@@ -4,7 +4,6 @@ import { Card, Button, SecondaryButton } from '../components/Common.js'
 import Accordion from '../components/Accordion'
 import Countdown from '../containers/Countdown'
 import Livestream from '../components/Livestream'
-import Schedule from '../components/Schedule'
 
 export default () => (
   <>
@@ -30,7 +29,5 @@ export default () => (
     </Accordion>
     <H2>Livestream Component</H2>
     <Livestream />
-    <H2>Schedule Component</H2>
-    <Schedule />
   </>
 )
